add react component
