+++ app/javascript/application.js
@@ -1,26 +1,17 @@
 // Entry point for the build script in your package.json
 import "@hotwired/turbo-rails";
 import "./controllers";
+
 import React from "react";
 import ReactDOM from "react-dom";
-import { Routes, Route } from "react-router-dom";
-import { Greetings } from "./components/greetings";
 import { Provider } from "react-redux";
-import store from "./redux/store";
+import App from "./App";
+import store from "./redux/configureStore";
 
-function App() {
-  return (
-    <Routes>
-      <Route index element={<Greetings />} />
-    </Routes>
-  );
-}
+const root = ReactDOM.createRoot(document.getElementById("root"));
 
-ReactDOM.render(
+root.render(
   <Provider store={store}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </Provider>,
-  document.getElementById("root")
+    <App />
+  </Provider>
 );
