+++ app/javascript/redux/configureStore.js
@@ -1,0 +1,10 @@
+import { configureStore } from "@reduxjs/toolkit";
+import messages from "./messages";
+
+const store = configureStore({
+  reducer: {
+    messages,
+  },
+});
+
+export default store;
