+++ app/javascript/redux/messages.js
@@ -1,6 +1,7 @@
+/* eslint-disable no-param-reassign */
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 
-const messagesAPI = "http://localhost:3000/api/v1/greeting";
+const messagesAPI = "http://localhost:3000/api/v1/messages/";
 const initialState = [];
 
 export const fetchMessage = createAsyncThunk(
@@ -8,6 +9,7 @@ export const fetchMessage = createAsyncThunk(
   async () => {
     const res = await fetch(messagesAPI);
     const data = await res.json();
+
     return data;
   }
 );
