+++ package.json
@@ -8,6 +8,7 @@
     "axios": "^1.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.4",
     "react-router-dom": "^6.4.2",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0"
